+++ lib/src/key_pair/ed25519_key_pair.dart
@@ -4,7 +4,6 @@ import 'package:cryptography/cryptography.dart' as crypto;
 import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
 import 'package:x25519/x25519.dart' as x25519;
 
-import '../digest_utils.dart';
 import '../exceptions/ssi_exception.dart';
 import '../exceptions/ssi_exception_type.dart';
 import '../types.dart';
@@ -93,17 +92,8 @@ class Ed25519KeyPair implements KeyPair {
       );
     }
 
-    Uint8List payloadToSign;
-    if (signatureScheme == SignatureScheme.eddsa_sha512) {
-      payloadToSign = data;
-    } else {
-      payloadToSign = DigestUtils.getDigest(
-        data,
-        hashingAlgorithm: signatureScheme.hashingAlgorithm,
-      );
-    }
-
-    return ed.sign(_privateKey, payloadToSign);
+    // For Ed25519, the library handles hashing internally
+    return ed.sign(_privateKey, data);
   }
 
   /// Verifies a signature using Ed25519.
@@ -132,16 +122,8 @@ class Ed25519KeyPair implements KeyPair {
       );
     }
 
-    Uint8List payloadToVerify;
-    if (signatureScheme == SignatureScheme.eddsa_sha512) {
-      payloadToVerify = data;
-    } else {
-      payloadToVerify = DigestUtils.getDigest(
-        data,
-        hashingAlgorithm: signatureScheme.hashingAlgorithm,
-      );
-    }
-    return ed.verify(ed.public(_privateKey), payloadToVerify, signature);
+    // For Ed25519, the library handles hashing internally
+    return ed.verify(ed.public(_privateKey), data, signature);
   }
 
   /// Returns the original seed used to derive the Ed25519 key pair.